Treated Duplicate Code Smell (Dispensable)
Noted that both diamond drawing code had similar functionality drawing the lower portion of the diamond. Since its practically drawing an upside down isosceles triangle, i extracted it as a method and named it as such.

However, since the drawing aspect of these methods revolve around having the appropriate spacing and number of stars to output, i also passed the correct state of said variables to the method.

Down the line, maybe a parameter object might be useful.